Fix feed path on blog home page
We've customised this to match WordPress's location but it's not picked
up by Minima.

I tried seeing if I could contribute this back but it was too hard.
`site.feed.path` is only available if the developer has customised that.
The jekyll-feed plugin checks for this or defaults to the root location.
We'd either need to re-implement this logic (nasty) or also update
jekyll-feed to always expose that configured setting (much nicer but now
I have Two Problems).