Fix code highlighting in posts
In WordPress I was using the CodeColorer plugin. This used a WordPress
shortcode to display syntax-highlighted code blocks. This doesn't work
in Jekyll and it has native support for code blocks with highlighting.

This change converts posts that had code markup to Markdown so that we
can use this build-in support for code blocks and highlighting.

I've cleaned up and converted a bit of the HTML in these post to
Markdown while I was at it.

Fixes #3.